requirements.txt
sqlalchemy
kafka-python
psycopg2